Add files via upload
Starting to add the hmdb search tool to the GUI. Now should open a new tab to search the database. Once database is searched you can select a compound to add. The next step will be to make it possible to add this data to the database. Will need to create a tool similar to the already existing pubchem add.