Add files via upload
Added a new feature to write chemicals to a personal chemical database after searching from the PMF database. This will allow for quicker access of necessary chemicals. Next step will be to make it create a new excel sheet from within the app if one currently doesn't exist.
Final search step will be in making a beautiful output for any selected compounds. This may include charts and images that can assist the researchers.