Add files via upload
Now will have the option to display more information about the compound after searching. This is important for advanced searches as it will help allow the manual narrowing of search results. I will probably make this so that these narrowed chemicals can be added to an individual's database for easier viewing and export.